+++ entrypoint.sh
@@ -7,6 +7,7 @@ file=$1
 provider=$2
 env=$3
 version=$4
+timeout=${5:-30}
 
 # Replace version placeholder with actual version
 jq --arg a $version '.Version = $a' $file > "cmr/cmr.json"
@@ -17,13 +18,13 @@ if [[ $env == "sit" || $env == "uat" ]]; then
 fi
 
 if [[ "${LAUNCHPAD_TOKEN_SIT}" ]] && [[ $env == "sit" ]]; then
-  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p POCLOUD -e ${env} -t $LAUNCHPAD_TOKEN_SIT
+  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p ${provider} -e ${env} -t $LAUNCHPAD_TOKEN_SIT -to $timeout
 elif [[ "${LAUNCHPAD_TOKEN_UAT}" ]] && [[ $env == "uat" ]]; then
-  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p POCLOUD -e ${env} -t $LAUNCHPAD_TOKEN_UAT
+  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p ${provider} -e ${env} -t $LAUNCHPAD_TOKEN_UAT -to $timeout
 elif [[ "${LAUNCHPAD_TOKEN_OPS}" ]] && [[ $env == "ops" ]]; then
-  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p POCLOUD -e ${env} -t $LAUNCHPAD_TOKEN_OPS
+  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p ${provider} -e ${env} -t $LAUNCHPAD_TOKEN_OPS -to $timeout
 else
-  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p POCLOUD -e ${env} -cu $cmr_user -cp $cmr_pass
+  umms_updater -d -f cmr/cmr.json -a cmr/${env}_associations.txt -p ${provider} -e ${env} -cu $cmr_user -cp $cmr_pass -to $timeout
 fi
 
 set -x